feat: refactor & implement connectable
Refactors and separates the components.
Opens the popover only when content is connected to DOM.
