+++ src/cosmoz-dropdown-menu.ts
@@ -1,0 +1,60 @@
+import { css, component } from '@pionjs/pion';
+import { html } from 'lit-html';
+import { Props as DropdownProps } from './cosmoz-dropdown';
+
+const style = css`
+	:host {
+		display: contents;
+		max-height: var(--cosmoz-dropdown-menu-max-height, calc(96vh - 64px));
+		overflow-y: auto;
+	}
+	::slotted(:not(slot)) {
+		display: block;
+		--paper-button_-_display: block;
+		box-sizing: border-box;
+		padding: 10px 24px;
+		background: transparent;
+		color: var(--cosmoz-dropdown-menu-color, #101010);
+		transition:
+			background 0.25s,
+			color 0.25s;
+		border: none;
+		cursor: pointer;
+		font-size: 14px;
+		line-height: 20px;
+		text-align: left;
+		margin: 0;
+		width: 100%;
+	}
+
+	::slotted(:not(slot):hover) {
+		background: var(
+			--cosmoz-dropdown-menu-hover-color,
+			var(--cosmoz-selection-color, rgba(58, 145, 226, 0.1))
+		);
+	}
+
+	::slotted(:not(slot)[disabled]) {
+		opacity: 0.5;
+		pointer-events: none;
+	}
+`;
+
+const List = () => html` <slot></slot> `;
+customElements.define(
+	'cosmoz-dropdown-list',
+	component(List, { styleSheets: [style] }),
+);
+
+type MenuProps = Pick<DropdownProps, 'placement'>;
+const Menu = ({ placement }: MenuProps) =>
+	html` <cosmoz-dropdown
+		.placement=${placement}
+		part="dropdown"
+		exportparts="anchor, button, content, wrap, dropdown"
+	>
+		<slot name="button" slot="button"></slot>
+		<cosmoz-dropdown-list><slot></slot></cosmoz-dropdown-list>
+	</cosmoz-dropdown>`;
+
+customElements.define('cosmoz-dropdown-menu', component<MenuProps>(Menu));
