+++ src/index.ts
@@ -1,3 +1,4 @@
-export * from './cosmoz-dropdown';
 export * from './use-focus';
 export * from './use-position';
+export * from './cosmoz-dropdown-menu';
+export * from './cosmoz-dropdown';
